### 2023-07-07 #5626 @editorialbot
[PRE REVIEW]: HeuDiConv — flexible DICOM conversion into structured directory layouts
**Submitting author:** @yarikoptic (<a href="http://orcid.org/0000-0003-3456-2493">Yaroslav Halchenko</a>)
**Repository:** https://github.com/nipy/heudiconv/
**Branch with paper.md** (empty if default branch): paper-joss
**Version:** 0.13.1
