### 2023-07-07 #5626 @Kevin-Mattheus-Moerman
I am the AEiC for JOSS on this topic track. I have just triggered a scope review for this submission by the editorial board. It is likely this work will not be deemed in in scope for JOSS due to its small size and utility tool like functionality (see also: https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). The board will review this work now for scope and will get back with a verdict hopefully within 2 weeks.
