### 2023-07-07 #5626 @yarikoptic
Dear @Kevin-Mattheus-Moerman . Please let us know if any extra information is required.

As far as "substantial" goes -- it is great to see that it qualified as "small" since frankly I do get lost in that codebase (over 10K LoC) at times. Code goes through a number of aspects in neuroinformatics - different file formats, metadata handling, harmonization and standardization. While working on those it triggered various developments and fixes to the other underlying tools it uses (e.g. `dcm2niix`) and standards (e.g. BIDS) themselves, and codebase possesses various someone unique and someone could even call "scientific" aspects such as determination of applicability of fieldmap correction maps etc.  Meanwhile let me also just go through the "substantial scholarly effort" items since I believe it overwhelmingly "qualifies" according to those:

- Age of software (is this a well-established software project) / length of commit history.

First commit: Wed Jan 29 19:20:12 2014

- Number of commits.

```shell
❯ git log --pretty=oneline | nl | wc -l
1573
```
- Number of authors.

```shell
❯ git shortlog -sn | grep -v auto | wc -l
47
```

- Total lines of code (LOC). Submissions under 1000 LOC will usually be flagged, those under 300 LOC will be desk rejected.

```shell
❯ git grep -v '^ *$' | wc -l
11732
```
- Whether the software has already been cited in academic papers.

yes.  some could be seen at https://scicrunch.org/resolver/SCR_017427/mentions?q=&i=rrid:scr_017427

- Whether the software is sufficiently useful that it is likely to be cited by your peer group.

see above.
