### 2023-07-07 #5626 @yarikoptic
as for precedents, I see JOSS papers (I will not point fingers but can cite if asked) which are really about much smaller  niche tools, in particular for conversion.
