### 2023-07-07 #5626 @yarikoptic
I think I got the reason for the concern @Kevin-Mattheus-Moerman  -- the "software report" from the bot running  `cloc` above is about "paper branch" (`paper-joss`) which is detached from the main codebase entirely - similarly to how we did for DataLad publication. And that is why it gives "wrong" counts -- it does not only contain paper but also some python code to collect co-authors info and email them...  Let me know if I should merge "master" branch into it, but here are the actual stats for the `master` branch with code (gives less than my simple grep assesment above -- forgot that we could have doclines etc ;)):

```shell
❯ git clone https://github.com/nipy/heudiconv
Cloning into 'heudiconv'...
remote: Enumerating objects: 7872, done.
remote: Total 7872 (delta 563), reused 905 (delta 554), pack-reused 6905
Resolving deltas: 100% (5166/5166), done.
❯ cd heudiconv
CHANGELOG.md  CONTRIBUTING.rst  Dockerfile  LICENSE  NOTES  README.rst  custom/  dev-requirements.txt  docs/  heudiconv/  mypy.ini  pyproject.toml  requirements.txt  setup.py*  tox.ini  utils/
❯ /home/yoh/proj/misc/cloc/cloc .
      92 text files.
      84 unique files.
      20 files ignored.

Language                             files          blank        comment           code
Python                                  43           1482           2492           7323
Markdown                                 2            189             12            555
reStructuredText                        15            173            104            359
YAML                                     9             45             10            265
Dockerfile                               1              2             11            145
Bourne Shell                             3             27             30             91
INI                                      2             15              0             89
Windows Module Definition                1              9              0             55
Bourne Again Shell                       2             12             16             41
Text                                     4              0              0             13
TOML                                     1              1              0              9
make                                     1              4              6              9
SUM:                                    84           1959           2681           8954
```
