### 2023-07-07 #5626 @Kevin-Mattheus-Moerman
@yarikoptic my apologies. I missed that. We usually detect that and do the line counts on the correct branch. Thanks for pointing this out. I'll remove the query-scope label.
