### 2023-07-07 #5626 @Kevin-Mattheus-Moerman
@yarikoptic note the paper error mentions `failure: Author (Krzysztof Jacek Gorgolewski) is missing affiliation`
