### 2023-07-07 #5626 @Kevin-Mattheus-Moerman
@yarikoptic I have just `waitlisted`, this is because all the editors in this domain are currently handling many other submissions. We will assign an editor as soon as one becomes available.

In the meantime you can address that paper error :point_up:, you may also use `@editorialbot check references` to see if there are any potentially missing/invalid DOIs, and then fix those too if needed.

You may also start suggesting reviewers here if you like. Please mention their GitHub handles but leave out the `@` symbol so they are not tagged.
