### 2023-07-07 #5626 @yarikoptic
> @yarikoptic note the paper error mentions `failure: Author (Krzysztof Jacek Gorgolewski) is missing affiliation`

I noted that too, although doesn't happen for me locally -- I guess we build differently.  The question is: does everybody have to have an affiliation? What if someone does work "on the side" and the work has nothing to do with the main day job (so not really "Independent Researcher" as in the [example](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography))?

In this case I believe it was also a conscious choice of @chrisgorgo (please clarify if I am wrong) since his current affiliation has nothing to do with this work, and he is no longer affiliated with an institution where/when this work was relevant in the past.

Anyways -- to resolve, I am adding  "Emeritus of Department of Psychology, Stanford University, CA, USA" to be most factually correct as in relation to this work.
