### 2023-07-12 #5626 @britta-wstnr
Hi @Kevin-Mattheus-Moerman - yes, this is something I could edit!
I assume it's clearer in the system if this is assigned to an editor already - if so, please go ahead. I imagine that as soon as I've found reviewers for my other papers, I could start looking at this one, too. (Note: I will be on leave for a few days starting next week though.)
