### 2023-07-12 #5626 @yarikoptic
Thank you @britta-wstnr for taking on editing our manuscript!

> You may also start suggesting reviewers here if you like.

I find it difficult to suggest anyone in particular given a notable list of COIs I and an extended list of co-authors would have -- anyone we would think of knowing would likely conflict  ;-)  @britta-wstnr would you be able to identify some willing hands among the pull of reviewers within JOSS's "Social, Behavioral, and Cognitive Sciences" category or would still prefer us to figure out some recommendations?
