### 2023-07-20 c640ffe
fix the order of args in the custom_seqinfo doc
