### 2023-07-20 e8e6a7d
fix the order of args in the custom_seqinfo doc
