### 2023-07-20 7904d46
ENH: raise exception if dcm2niix returns with non-0 code
