### 2023-07-20 b7f1bf2
Prototype steps for where to add handling of renaming after dcm2niix
