### 2023-08-30 2a1e7b4
Boost claimed BIDS version to 1.8.0 from 1.4.1
AFAIK there were no relevant to us breakages. So just boosting the version
to seemingly be up to date with BIDS.

Closes #698
