+++ heudiconv/bids.py
@@ -68,7 +68,7 @@ class BIDSError(Exception):
     pass
 
 
-BIDS_VERSION = "1.4.1"
+BIDS_VERSION = "1.8.0"
 
 # List defining allowed parameter matching for fmap assignment:
 SHIM_KEY = "ShimSetting"
