### 2023-08-30 38a80b5
Initial sketch for implementing +? which would make increment within that sequence
request by @dnkennedy
