### 2023-09-01 fc22c92
ENH: flake8 - use kwarg maxsplit in re.split
