+++ .github/workflows/test.yml
@@ -21,7 +21,7 @@ jobs:
           - '3.11'
     steps:
       - name: Check out repository
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
         with:
           fetch-depth: 0
 
