+++ tox.ini
@@ -8,12 +8,14 @@ commands = pytest -s -v {posargs} heudiconv
 [testenv:lint]
 skip_install = True
 deps =
+    codespell
     flake8
     flake8-bugbear
     flake8-builtins
     flake8-unused-arguments
 commands =
     flake8 heudiconv
+    codespell
 
 [testenv:typing]
 deps =
