### 2023-09-28 330ed76
Convert assertion into a warning that we would not use dicom dir template option
To avoid users confusion etc. Ref: https://github.com/nipy/heudiconv/issues/707#issuecomment-1738371656
