### 2023-10-06 cdb74df
added elements required for asl
