### 2023-10-09 0f8c14b
added asl-support content in utils.py and dicoms.py files
