+++ heudiconv/dicoms.py
@@ -109,6 +109,11 @@ def create_seqinfo(mw: dw.Wrapper, series_files: list[str], series_id: str) -> S
         date=dcminfo.get("AcquisitionDate"),
         series_uid=dcminfo.get("SeriesInstanceUID"),
         time=dcminfo.get("AcquisitionTime"),
+        # ASL support
+        in_plane_phase_encoding_direction = dcminfo.get("InPlanePhaseEncodingDirection") if dcminfo.get("InPlanePhaseEncodingDirection") else None,
+        acquisition_number = dcminfo.get("AcquisitionNumber") if dcminfo.get("AcquisitionNumber") else None,
+        image_orientation_patient = tuple(dcminfo.get('ImageOrientationPatient')) if dcminfo.get('ImageOrientationPatient') else ()
+
     )
 
 
