### 2023-10-11 #5839 @marcelzwiers
Heudiconv is probably the best known and most used software tool for researchers who would like to do open science and organize their data according to the BIDS standard. The heudiconv paper is well written but brief and rather incomplete without the online read-the-docs documentation. I felt that after reading the paper I didn't get an idea about the architecture of Heudiconv, i.e. how the heuristics are used, what the main workflow is (the paper mentions "discovery, manual tuning, conversion" steps without properly explaining what they are. In the tutorials it is mentioned that you first may have to generate a `dicominfo.tsv` file that can then be used to construct a heuristic file, but as a reader you have to connect all the dots yourself. To get a better understanding of the heuristics file, I looked at a few (third party) tutorials and opened the `studyforrest_phase2.py` file. Then I understood a bit more, but many things remained unexplained or untouched. I then opened the `reproin.py` file, but found a staggering 926 lines of code to reverse engineer... I understand that this is advanced usage that is beyond the scope of the paper, but I would expect a basic explanation of the workflow of getting a `dicominfo.tsv` file, an explanation that you, as a user, have to write python code to make conversion `keys` from `seqinfo` data that is provided by heudiconv (a little bit of heuristics code and a snippet of seqinfo would have been helpful).

About the paper, I have a few more comments:

Figure 1 contains details about the data acquisition (k-space!) that have nothing to do with heudiconv. Also, the figure contains several post-heudiconv use cases, which are very general and uninformative about heudiconv. In the figure, heudiconv itself is just one black box! Altogether, I suggest to remove the figure or replace it with a flow-chart of heudiconv itself (the inside of the black box), or of it's workflow (or both).

Open general checks:
- Data sharing: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.
- Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.

Figure 2 from the paper contains original data (arguably) but I cannot access the data nor reproduce the figure. I looked into etelemetry but the data is stored on a private server. I also looked at the pypi data, but in order to access that I believe I would need a (payed) Google Cloud BigQuery account. I am fine with accepting these checks if the editor does not consider download metrics as original data (although it would be nice if the authors explain a bit more on how the data was generated)

Open functionality checks:
-Functionality: Have the functional claims of the software been confirmed?

The paper makes many functional claims, e.g. that it can handle complex data structures, data acquisitions schemes, and much more. I could not verify all claims but did try to convert a (relatively simple) standard dataset using the `convertall` heuristics and `--bids` flag. That worked (after some struggles) but the output was not in BIDS format. Yet, I do suspect that this wouldn't have been the case if I had written a custom heuristics file. And given the evidence of researchers in the field having used heudiconv to convert more complex datasets to BIDS I'm happy to check this box if the editor is ok with that.

Open documentation checks:
- Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
- Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

I have already touched upon these checks in my general comment. Several examples (if third party tutorials count as examples) can be found on the read-the-docs website, but not in the paper. I would have liked to see a simple example in the paper (not in detail, but to get an idea what of what is involved and what the user would need to do). The same is true for the core functionality, it is described at the website but the paper basically only states "that decision making is implemented in HeuDiConv heuristics" and that heudiconv can operate in BIDS mode with a `--bids` flag.

Open software paper checks:
- State of the field: Do the authors describe how this software compares to other commonly-used packages?

Importantly, this is missing entirely. The limitations of heudiconv are not discussed (I think it should at least be mentioned that you have to be able to program Python, get to know the heudiconv API + that you need to have some understanding of pulse sequence parameters), neither is there discussion how heudiconv's advantages and disadvantages compare to those of other tools in the field.
