### 2023-10-19 #5839 @britta-wstnr
Hi @marcelzwiers -
thank you very much for your detailed review! 👍
One comment for you and @yarikoptic: your feedback towards the gaps in the paper touches upon a specialty of JOSS papers. They are meant to be very brief (250 to 1000 words only! https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) and should not contain documentation of API* etc (https://joss.readthedocs.io/en/latest/review_criteria.html).
However, this _should_ be part of the documentation, so if you @marcelzwiers had problems understanding the architecture/API of the toolbox based on the docs, then this should be changed.

*one reason for this is that API documentation in papers often becomes obsolete after a few years, while the docs are a "living" thing that changes alongside the code.
