### 2023-10-30 #5839 @JLefortBesnard
Hello,
Setting up the process of converting DICOM to BIDS is not an easy task, and yet, very much necessary for reproducibility. HeuDiConv, as a software tool to provide advanced and adaptable BIDS dataset creation, fully meets this request and the range of available pre-defined heuristics simplifies the usage.

Many great tutorials are available, I followed two of them and found them quite clear. In fact, the tutorials' quality accentuates the disparity with the "read-the-docs" even more. Though it is sufficiently clear to individuals who have prior experience in converting Dicom to Bids using Python (that is why I did check the 'functionality documentation' box), I feel that the documentation is perhaps not designed in the most user-friendly manner for newcomers, and thus does not provide enough guidance to make the usage more intuitive. Striving in this direction might be beneficial in promoting the use of this tool.

I find the paper well-crafted and providing enough information for a solid understanding of the software. I appreciate Figure 1, which effectively illustrates the software's functionality. However, I have reservations about Figure 2, which, in my opinion, may not be relevant to the user since the information can be adequately summarized in the text. Instead, including a second figure that depicts the inner workings of the software, as the first reviewer suggested, would be more engaging.

I also agree with the first reviewer about the missing limitations of HeuDiConv and the missing comparison with other similar tools (or ways) to convert DICOM into Bids (that is why I did not check the 'state of the field' box).

I did not test all functional claims made in the paper ("complex organization scenario")  thus cannot confirm them.

Overall, I think this paper belongs in JOSS. This tool is poised to make a significant impact on the neuroimaging community, and I will use it too.
