### 2023-10-30 nipy/heudiconv#715 @yarikoptic
[JOSS]: Add more on state of the field
Both reviewers raised concern in https://github.com/openjournals/joss-reviews/issues/5839
