### 2023-10-30 nipy/heudiconv#716 @yarikoptic
[JOSS]: Add a figure outlining typical heudiconv "workflow"
In response to https://github.com/openjournals/joss-reviews/issues/5839 , both reviewers had expressed difficulty orienting themselves in how to use heudiconv.  Reviewer 1 also did not appreciate Figure 1 IMHO primarily because there were no figure showing internal heudiconv workflow.  So if we add a figure outlining stages of the heudiconv work, may be on 1 simple example (`convertall`?) -- it could help.

- @yarikoptic : verbalize the example of two stage and one stage modes of operation.
- @TheChymera : draw a figure, alike to  https://fw-heudiconv.readthedocs.io/en/latest/overview.html#general-workflow
