### 2023-11-02 #5839 @JLefortBesnard
> Thank you @JLefortBesnard for the review! Would you mind sharing which particular 2 tutorials you found "quite clear"?

Sure, I enjoyed the tutorials from James Kent and the one from Stanford
