### 2023-11-20 24127ff
ENH: give informative assertion message when multiple values are found
