### 2023-12-08 9e254c8
Drop Python 3.7 and announce support for 3.12
