+++ .github/workflows/test.yml
@@ -14,11 +14,11 @@ jobs:
       fail-fast: false
       matrix:
         python-version:
-          - '3.7'
           - '3.8'
           - '3.9'
           - '3.10'
           - '3.11'
+          - '3.12'
     steps:
       - name: Check out repository
         uses: actions/checkout@v4
