### 2023-12-08 d876968
[DATALAD RUNCMD] Use 3.8 in workflows
=== Do not change lines below ===
{
 "chain": [],
 "cmd": "sed -i -e 's,3\\.7,3.8,g' .github/workflows/codespell.yml .github/workflows/docker.yml .github/workflows/lint.yml .github/workflows/release.yml .github/workflows/test.yml .github/workflows/typing.yml",
 "exit": 0,
 "extra_inputs": [],
 "inputs": [],
 "outputs": [],
 "pwd": "."
}
^^^ Do not change lines above ^^^
