+++ .github/workflows/release.yml
@@ -36,7 +36,7 @@ jobs:
         if: steps.auto-version.outputs.version != ''
         uses: actions/setup-python@v4
         with:
-          python-version: '^3.7'
+          python-version: '^3.8'
 
       - name: Install Python dependencies
         if: steps.auto-version.outputs.version != ''
