### 2023-12-08 2e761a5
Not ready for 3.12 yet due to pylibjpeg-libjpeg/
