+++ .github/workflows/test.yml
@@ -18,7 +18,6 @@ jobs:
           - '3.9'
           - '3.10'
           - '3.11'
-          - '3.12'
     steps:
       - name: Check out repository
         uses: actions/checkout@v4
