### 2024-01-04 17c4968
New wiorkflow figure and caption, renamed figures, better style consistency
