+++ scripts/environment.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+dot -Tpdf ../data/environment.dot -o ../figs/environment.pdf
