### 2024-01-05 a264f28
New figure style and PNG outputs as per yarik
