### 2024-01-05 4c601b2
Uncommented PNG generation
