### 2024-01-12 585a922
Fix typos in added figure caption
