### 2024-01-12 2d3c284
Adjust legend for heudiconv 2 ways figure 2
