+++ paper.md
@@ -186,7 +186,7 @@ It also ensures files are named according to the BIDS specifications, including 
 ![**HeuDiConv automates the keystone conversion step in reproducible data handling, without compromising operator flexibility.** The showcased set-up depicts a 2-machine infrastructure, with heudiconv operating on the same machine as subsequent analysis steps for data in a standardized and shareable representation. For more advanced usage at institutions with dedicated infrastructure, HeuDiConv can operate on an additional third machine, interfacing between the depicted two, and dedicated to data repositing, versioning, and backup.](figs/environment.pdf)
 
 
-![**Heudicom conversion and layout can be controlled via custom heuristiscs or by selecting built-in presets.** The heudicom application can be used with the `-c none` parameter to generate a heuristic template for the user to edit, before conversion is performed e.g. via the `-c dcm2niix` parameter which uses the eponymous external application. The process is itempotent, and specifying the `-c dcm2niix` parameter will automatically convert based on a default layout and produce a heuristic directory for the user to inspect post hoc.](figs/workflow.pdf)
+![**HeuDiConv conversion and layout is controlled via heuristics (custom or provided built-ins) either with manual tune up of proposed filenames or fully automated.** The heudiconv application can be used with the `-c none` parameter to generate by heuristic a list of filenames for the user to edit, before invoking the conversion to be performed via the `-c dcm2niix` option to use the `dcm2niix` tool. The process is idempotent, and specifying the `-c dcm2niix` option can automatically convert without seeking user tune up of proposed filenames.](figs/workflow.pdf)
 
 ## Exemplar heuristics
 
