### 2024-01-12 d89bf6e
Adding workflow figure
