### 2024-01-12 3aa4b2d
Add documentation building instructions
