### 2024-01-12 65a05c8
rst paths must be relative to the docs/ dir
