### 2024-01-16 65b5fb2
change tostr -> tobytes for deprecation, add basic test
