+++ heudiconv/convert.py
@@ -223,7 +223,7 @@ def prep_conversion(
                 custom_grouping=getattr(heuristic, "grouping", None),
                 # callable which will be provided dcminfo and returned
                 # structure extend seqinfo
-                custom_seqinfo = getattr(heuristic, 'custom_seqinfo', None),
+                custom_seqinfo=getattr(heuristic, 'custom_seqinfo', None),
             )
         elif seqinfo is None:
             raise ValueError("Neither 'dicoms' nor 'seqinfo' is given")
