+++ heudiconv/dicoms.py
@@ -9,7 +9,18 @@ import os.path as op
 from pathlib import Path
 import sys
 import tarfile
-from typing import TYPE_CHECKING, Any, Dict, Hashable, List, NamedTuple, Optional, Union, Protocol, overload
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    Dict,
+    Hashable,
+    List,
+    NamedTuple,
+    Optional,
+    Protocol,
+    Union,
+    overload,
+)
 from unittest.mock import patch
 import warnings
 
@@ -90,14 +101,19 @@ def create_seqinfo(
     global total_files
     total_files += len(series_files)
 
-    custom_seqinfo_data = custom_seqinfo(wrapper=mw, series_files=series_files) \
-        if custom_seqinfo else None
+    custom_seqinfo_data = (
+        custom_seqinfo(wrapper=mw, series_files=series_files)
+        if custom_seqinfo
+        else None
+    )
     try:
         hash(custom_seqinfo_data)
     except TypeError:
-        raise RuntimeError("Data returned by the heuristics custom_seqinfo is not hashable. "
-                           "See https://heudiconv.readthedocs.io/en/latest/heuristics.html#custom_seqinfo for more "
-                           "details.")
+        raise RuntimeError(
+            "Data returned by the heuristics custom_seqinfo is not hashable. "
+            "See https://heudiconv.readthedocs.io/en/latest/heuristics.html#custom_seqinfo for more "
+            "details."
+        )
 
     return SeqInfo(
         total_files_till_now=total_files,
