### 2024-01-18 35ad528
Copy tutorials from University of Arizona
Originally written by Dianne Patterson for the U of A group, these
tutorials are generally useful to all heudiconv users.

This commit brings those tutorials verbatim and will be edited in
subsequent commits.

    Co-authored-by: Dianne Patterson <dkp@arizona.edu>
