+++ docs/tutorials.rst
@@ -2,6 +2,504 @@
 User Tutorials
 ==============
 
+
+Lesson 1: Running heuristic.py
+*********************************
+
+* Download and unzip `sub-219_dicom.zip <https://osf.io/mqgzh/>`_. You will see a directory called MRIS.
+
+* Under the MRIS directory, is the *dicom* subdirectory: Under the subject number *219* the session *itbs* is nested.  Each dicom sequence folder is nested under the session.  You can delete sequences folders if they are of no interest::
+
+    dicom
+    └── 219
+        └── itbs
+            ├── Bzero_verify_PA_17
+            ├── DTI_30_DIRs_AP_15
+            ├── Localizers_1
+            ├── MoCoSeries_19
+            ├── MoCoSeries_31
+            ├── Post_TMS_restingstate_30
+            ├── T1_mprage_1mm_13
+            ├── field_mapping_20
+            ├── field_mapping_21
+            └── restingstate_18
+
+
+* Pull the HeuDiConv Docker container to your machine::
+
+    docker pull nipy/heudiconv
+
+* From a BASH shell (no, zsh will not do), navigate to the MRIS directory and run the ``hdc_run.sh`` script for subject *219*, session *itbs*, like this::
+
+  ./hdc_run.sh heuristic1.py 219 itbs
+
+* This should complete the conversion. After running, the *Nifti* directory will contain a bids-compliant subject directory::
+
+
+    └── sub-219
+        └── ses-itbs
+            ├── anat
+            ├── dwi
+            ├── fmap
+            └── func
+
+* The following required BIDS text files are also created in the Nifti directory. Details for filling in these skeleton text files can be found under `tabular files <https://bids-specification.readthedocs.io/en/stable/02-common-principles.html#tabular-files>`_ in the BIDS specification::
+
+    CHANGES
+    README
+    dataset_description.json
+    participants.json
+    participants.tsv
+    task-rest_bold.json
+
+* Next, visit the `bids validator <https://bids-standard.github.io/bids-validator/>`_.
+* Click `Choose File` and then select the *Nifti* directory.  There should be no errors (though there are a couple of warnings).
+
+  .. Note:: Your files are not uploaded to the BIDS validator, so there are no privacy concerns!
+* Look at the directory structure and files that were generated.
+* When you are ready, remove everything that was just created::
+
+    rm -rf Nifti/sub-* Nifti/.heudiconv Nifti/code/__pycache__ Nifti/*.json Nifti/*.tsv Nifti/README Nifti/CHANGE
+
+* Now you know what the results should look like.
+* In the following sections, you will build *heuristic.py* yourself so you can test different options and understand how to work with your own data.
+
+
+Running HeuDiConv is a 3-step process
+*********************************************
+
+* :ref:`Step1 <heudiconv_step1>` By passing some path information and flags to HeuDiConv, you generate a heuristic (translation) file skeleton and some associated descriptor text files.  These all get placed in a hidden directory, *.heudiconv* under the *Nifti* directory.
+* :ref:`Step2 <heudiconv_step2>` Copy *MRIS/Nifti/.heudiconv/heuristic.py* to *MRIS/Nifti/code/heuristic.py*. You will modify the copied *heuristic.py* to specify BIDS output names and directories, and the input DICOM characteristics.  Available input DICOM characteristics are listed in *MRIS/Nifti/.heudiconv/dicominfo.tsv*.
[…]
