+++ docs/quickstart.rst
@@ -1,0 +1,125 @@
+Quickstart
+==========
+
+Heudiconv Hello World: Using `heuristic.py`
+
+.. TODO convert to a datalad dataset 
+.. TODO ``datalad install https://osf.io/mqgzh/``
+.. TODO delete any sequences of no interest prior to push, lets make the
+   example ds only contain what is needed for these tutorials
+.. TODO create a docker/podman section explaining how to use containers
+   in lieu of `heudiconv`, change the tutorials to `heudiconv`, not
+   container.
+.. TODO convert bash script to docs
+
+This section demonstrates how to use the heudiconv tool with `heuristic.py` to convert DICOMS into the BIDS data structure.
+
+* Download and unzip `sub-219_dicom.zip <https://osf.io/mqgzh/>`_. You will see a directory called MRIS.
+* Under the MRIS directory, is the *dicom* subdirectory: Under the subject number *219* the session *itbs* is nested.  Each dicom sequence folder is nested under the session.  You can delete sequences folders if they are of no interest::
+
+    dicom
+    └── 219
+        └── itbs
+            ├── Bzero_verify_PA_17
+            ├── DTI_30_DIRs_AP_15
+            ├── Localizers_1
+            ├── MoCoSeries_19
+            ├── MoCoSeries_31
+            ├── Post_TMS_restingstate_30
+            ├── T1_mprage_1mm_13
+            ├── field_mapping_20
+            ├── field_mapping_21
+            └── restingstate_18
+
+
+* Pull the HeuDiConv Docker container to your machine::
+
+    docker pull nipy/heudiconv
+
+* From a BASH shell (no, zsh will not do), navigate to the MRIS directory and run the ``hdc_run.sh`` script for subject *219*, session *itbs*, like this::
+
+    #!/bin/bash
+    
+    : <<COMMENTBLOCK
+    This code calls the docker heudiconv tool to convert DICOMS into the BIDS data structure.
+    It requires that you are in the parent directory of both the Dicom and Nifti directories 
+    AND that your Nifti directory contain a subdirectory called code with the conversion routine, e.g., heuristic.py in it. 
+    See https://neuroimaging-core-docs.readthedocs.io/en/latest/pages/heudiconv.html.
+    See also https://heudiconv.readthedocs.io/en/latest/
+    COMMENTBLOCK
+    
+    
+    # Exit if number of arguments is less than 3
+    if [ $# -lt 3 ]
+        then
+            echo "======================================================"
+            echo "Three arguments are required:"
+            echo "argument 1: name of conversion file in the Nifti/code directory, e.g., heuristic.py"
+            echo "argument 2: name of subject dicom folder to convert"
+            echo "argument 3: optional name of the session"
+            echo "e.g., $0 heuristic.py 219 itbs"
+            echo "output will be a BIDS directory under the Nifti folder"
+            echo "This assumes you are running docker"
+            echo "and have downloaded heudiconv: docker pull nipy/heudiconv"
+            echo "It also assumes that you are running from the parent directory to both dicom and Nifti"
+            echo "If you have a session argument, this assumes your DICOMS are nested under subject and then session"
+            echo "Finally, note that the dicoms are assumed to be *.dcm files"
+            echo "======================================================"
+            exit 1
+    fi
+    
+    # Define the three variables
+    converter=${1}
+    subject=${2}
+    session=${3}
+    
+    echo "Nifti/code/${converter} will be used to convert subject ${subject} and session ${session} under dicom"
+    echo "to BIDS output under Nifti/sub-${subject} and session ${session}"
+    
+    # This docker command assumes you are in in the bound (-v) base directory, e.g., the unzipped MRIS directory (PWD).
[…]
