### 2024-01-18 bd717de
Add links to original tutorials
