### 2024-01-22 5550deb
Minimize quickstart, format, and hand-test
