+++ docs/quickstart.rst
@@ -1,23 +1,23 @@
 Quickstart
 ==========
 
+This section demonstrates how to use the heudiconv tool with a provided `heuristic.py` to convert DICOMS into the BIDS data structure.
+
 This tutorial is based on https://neuroimaging-core-docs.readthedocs.io/en/latest/pages/heudiconv.html#lesson-1-running-heuristic-py
 
-Heudiconv Hello World: Using `heuristic.py`
+Install Prerequisites
+*********************
 
-.. TODO convert to a datalad dataset 
-.. TODO ``datalad install https://osf.io/mqgzh/``
-.. TODO delete any sequences of no interest prior to push, lets make the
-   example ds only contain what is needed for these tutorials
-.. TODO create a docker/podman section explaining how to use containers
-   in lieu of `heudiconv`, change the tutorials to `heudiconv`, not
-   container.
-.. TODO convert bash script to docs
+`dcm2niix` is the engine that will do the DICOM conversion, so make sure it is installed (or use a heudiconv container TODO link)::
 
-This section demonstrates how to use the heudiconv tool with `heuristic.py` to convert DICOMS into the BIDS data structure.
+    pip install dcm2niix
 
-* Download and unzip `sub-219_dicom.zip <https://osf.io/mqgzh/>`_. You will see a directory called MRIS.
-* Under the MRIS directory, is the *dicom* subdirectory: Under the subject number *219* the session *itbs* is nested.  Each dicom sequence folder is nested under the session.  You can delete sequences folders if they are of no interest::
+Prepare Dataset
+***************
+
+Download and unzip `sub-219_dicom.zip <https://osf.io/mqgzh/>`_. 
+
+We will be working from a directory called MRIS. Under the MRIS directory is the *dicom* subdirectory: Under the subject number *219* the session *itbs* is nested.  Each dicom sequence folder is nested under the session::
 
     dicom
     └── 219
@@ -32,96 +32,66 @@ This section demonstrates how to use the heudiconv tool with `heuristic.py` to c
             ├── field_mapping_20
             ├── field_mapping_21
             └── restingstate_18
+    Nifti
+    └── code 
+        └── heuristic1.py
+
+Basic Conversion
+****************
+
+Next we will use heudiconv convert DICOMS into the BIDS data structure.
+The example dataset includes an example heuristic file, `heuristic1.py`.
+Typical use of heudiconv will require the creation
+and/or editing of your heuristic file (TODO link), which we will cover
+in the later tutorials (TODO link).
+
+    .. note:: Heudiconv requires you to run the command from the parent
+              directory of both the Dicom and Nifti directories, which is `MRIS` in
+              our case.
+
+Run the following command::
+
+    heudiconv  --files dicom/219/itbs/*/*.dcm -o Nifti -f Nifti/code/heuristic1.py -s 219 -ss itbs -c dcm2niix -b --minmeta --overwrite
 
 
-* Pull the HeuDiConv Docker container to your machine::
+* We specify the dicom files to convert with `--files`
+* The heuristic file is provided with the `-f` option
+* We tell heudiconv to place our output in the Nifti dir with `-o`
+* `-b` indicates that we want to output in BIDS format
+* `--minmeta` guarantees that meta-information in the dcms does not get inserted into the JSON sidecar. This is good because the information is not needed but can overflow the JSON file causing some BIDS apps to crash.
 
-    docker pull nipy/heudiconv
-
-* From a BASH shell (no, zsh will not do), navigate to the MRIS directory and run the ``hdc_run.sh`` script for subject *219*, session *itbs*, like this::
-
-    #!/bin/bash
+Output
+******
     
-    : <<COMMENTBLOCK
[…]
