### 2024-01-23 f0251cc
First pass custom heuristics tutorial
