### 2024-01-23 b2f2f10
second naive pass of custom heuristic
