+++ docs/custom-heuristic.rst
@@ -2,7 +2,7 @@
 Custom Heuristics
 =========================
 
-In this tutorial we will:
+In this tutorial we go more in depth, creating our own *heuristic.py* and modifying it for our needs:
 
 1. :ref:`Step1 <heudiconv_step1>` Generate a heuristic (translation) file skeleton and some associated descriptor text files.
 2. :ref:`Step2 <heudiconv_step2>` Modify the *heuristic.py* to specify BIDS output names and directories, and the input DICOM characteristics.
@@ -22,7 +22,8 @@ Prerequisites
 Step 1: Generate Skeleton
 *************************
 
-.. note:: Step 1 only needs to be completed once correctly for each project.
+.. note:: Step 1 only needs to be completed once for each project.
+   If repeating this step, ensure that the .heudiconv directory is removed.
 
 From the *MRIS* directory, run the following command to process the ``dcm`` files that you downloaded and unzipped for this tutorial.::
 
@@ -56,8 +57,10 @@ Step 2: Modify Heuristic
 .. TODO Lets remove heuristic1 and heuristic2 and create a 2nd example
    dataset? or branch?
 
-* You will modify three sections in *heuristic.py*. It is okay to rename this file, or to have several versions with different names. You just don't want to mix up your new *heuristic.py* and the finished *heuristic1.py* while you are learning.
-* Your goal is to produce a working *heuristic.py* that will arrange the output in a BIDS directory structure. Once you create a working *heuristic.py*, you can run it for different subjects and sessions (keep reading).
+We will modify the generated *heuristic.py* so heudiconv will arrange the output in a BIDS directory structure.
+
+It is okay to rename this file, or to have several versions with different names, just be sure to pass the intended filename with `-f`. See :doc:`heuristics` docs for more info.
+
 * I provide three section labels (1, 1b and 2) to facilitate exposition here. Each of these sections should be manually modified by you for your project.
 
 Section 1
@@ -117,7 +120,7 @@ Section 1
   * Define the output directories and file names according to the `BIDS specification <https://bids-specification.readthedocs.io/en/stable/04-modality-specific-files/01-magnetic-resonance-imaging-data.html>`_
   * Note the output names for the fieldmap images (e.g., *sub-219_ses-itbs_dir-PA_epi.nii.gz*, *sub-219_ses-itbs_magnitude1.nii.gz*, *sub-219_ses-itbs_magnitude2.nii.gz*, *sub-219_ses-itbs_phasediff.nii.gz*).
   * The reverse_phase encode dwi image (e.g., *sub-219_ses-itbs_dir-PA_epi.nii.gz*) is grouped with the fieldmaps because it is used to correct other images.
-  * Data that is not yet defined in the BIDS specification will cause the bids-validator to produce an error unless you include it in a 
+  * Data that is not yet defined in the BIDS specification will cause the bids-validator to produce an error unless you include it in a
 
 .. TODO new link :ref:`.bidsignore <bidsignore>` file.
 
@@ -219,22 +222,21 @@ Section 2
 
 .. _heudiconv_step3:
 
-Step 3: 
+Step 3:
 *******************
 
 * You have now done all the hard work for your project. When you want to add a subject or session, you only need to run this third step for that subject or session (A record of each run is kept in .heudiconv for you)::
 
-    docker run --rm -it -v ${PWD}:/base nipy/heudiconv:latest -d /base/dicom/{subject}/*/*.dcm -o /base/Nifti/ -f /base/Nifti/code/heuristic.py -s 219 -ss itbs -c dcm2niix -b --minmeta --overwrite
[…]
