### 2024-01-23 18d5e68
Update heudiconv/convert.py
Co-authored-by: Yaroslav Halchenko <debian@onerussian.com>
