### 2024-01-23 bc4ce42
cast to list first, add unit test for bvals_are_zero
