### 2024-01-24 #5839 @yarikoptic
Thank you @britta-wstnr !  We are slowly addressing concerns and recommendations raised by the reviewers such as needing improvements to documentation etc eg
- checkout these tutorials we found and incorporating now!:  https://github.com/nipy/heudiconv/pull/734
- added to README.md (which also got improved) and the manuscript a figure illustrating typical "workflow" using heudiconv: https://github.com/nipy/heudiconv?tab=readme-ov-file#howto-101
- and I am yet to expand on the state of the field (issue https://github.com/nipy/heudiconv/issues/715)

and then we would ping back here&reviewers for re-review.
