### 2024-01-31 40b3880
Cleanup quickstart and use install link
