+++ docs/reproin.rst
@@ -2,47 +2,55 @@
 Reproin 
 ================
 
-This tutorial is based on https://neuroimaging-core-docs.readthedocs.io/en/latest/pages/heudiconv.html#lesson-3-reproin-py
+This tutorial is based on `Dianne Patterson's University of Arizona tutorials <https://neuroimaging-core-docs.readthedocs.io/en/latest/pages/heudiconv.html#lesson-3-reproin-py>`_
 
-If you don't want to modify a Python file as you did for *heuristic.py*, an alternative is to name your image sequences at the scanner using the *reproin* naming convention. Take some time getting the scanner protocol right, because it is the critical job for running *reproin*. Then a single Docker command converts your DICOMS to the BIDS data structure. There are more details about *reproin* in the 
-.. TODO new link ref:`Links <heudiconv_links>` section above.
+`Reproin <https://github.com/ReproNim/reproin>`_ is a setup for
+automatic generation of sharable, version-controlled BIDS datasets from
+MR scanners.
 
-* You should already have Docker installed and have downloaded HeuDiConv as described in Lesson 1.
-* Download and unzip the phantom dataset: `reproin_dicom.zip <https://osf.io/4jwk5/>`_ generated here at the University of Arizona on our Siemens Skyra 3T with Syngo MR VE11c software on 2018_02_08.
-* You should see a new directory *REPROIN*. This is a simple reproin-compliant dataset without sessions. Derived dwi images (ADC, FA etc.) that the scanner produced were removed.
-* Change directory to *REPROIN*. The directory structure should look like this::
+If can control how your image sequences are named at the scanner you can
+use the *reproin* naming convention.  
 
-    REPROIN
-    ├── data
-    └── dicom
-        └── 001
-            └── Patterson_Coben\ -\ 1
-                ├── Localizers_4
-                ├── anatT1w_acqMPRAGE_6
-                ├── dwi_dirAP_9
-                ├── fmap_acq4mm_7
-                ├── fmap_acq4mm_8
-                ├── fmap_dirPA_15
-                └── func_taskrest_16
+Get Example Dataset
+-------------------
 
-* From the *REPROIN* directory, run this Docker command::
+This example uses a phantom dataset: `reproin_dicom.zip <https://osf.io/4jwk5/>`_ generated by the University of Arizona on their Siemens Skyra 3T with Syngo MR VE11c software on 2018_02_08.
 
-    docker run --rm -it -v ${PWD}:/base nipy/heudiconv:latest -f reproin --bids  -o /base/data --files /base/dicom/001 --minmeta
-* ``--rm`` means Docker should cleanup after itself
-* ``-it`` means Docker should run interactively
-* ``-v ${PWD}:/base`` binds your current directory to ``/base`` inside the container.  Alternatively, you could provide an **absolute path** to the *REPROIN* directory.
-* ``nipy/heudiconv:latest`` identifies the Docker container to run (the latest version of heudiconv).
+The ``REPROIN`` directory is a simple reproin-compliant DICOM (.dcm) dataset without sessions. 
+(Derived dwi images (ADC, FA etc.) that the scanner produced have been removed.::
+
+    [user@local ~/reproin_dicom/REPROIN]$ tree -I "*.dcm"
+
+        REPROIN
+        ├── data
+        └── dicom
+            └── 001
+                └── Patterson_Coben\ -\ 1
+                    ├── Localizers_4
+                    ├── anatT1w_acqMPRAGE_6
+                    ├── dwi_dirAP_9
+                    ├── fmap_acq4mm_7
+                    ├── fmap_acq4mm_8
+                    ├── fmap_dirPA_15
+                    └── func_taskrest_16
+
[…]
