### 2024-02-01 688be6e
Document provided heuristics
