+++ docs/heuristics.rst
@@ -12,6 +12,14 @@ covered by the existing heuristics. This section will outline what makes up a
 heuristic file, and some useful functions available when making one.
 
 
+Provided Heuristics
+-------------------
+
+Running ``heudiconv`` without a heuristic file results in the generation of a skeleton for the user to customize to their needs. 
+
+``heudiconv`` also provides more than 10 additional heuristics, which can be seen `here <https://github.com/nipy/heudiconv/tree/master/heudiconv/heuristics>`_
+These heuristic files are documented in their code comments.
+
 Components
 ==========
 
