+++ docs/commandline.rst
@@ -10,11 +10,3 @@ paths.
    :prog: heudiconv
    :nodefault:
    :nodefaultconst:
-
-TODO
-====
-
-List provided heuristics
-
- e.g., banda-bids.py, bids_with_ses.py, cmrr_heuristic.py, example.py, multires_7Tbold.py, reproin.py, studyforrest_phase2.py, test_reproin.py, uc_bids.py. *heuristic.py* is a good default though.
-
