+++ .github/workflows/test.yml
@@ -53,7 +53,7 @@ jobs:
         run: coverage run `which pytest` -s -v heudiconv
 
       - name: Upload coverage to Codecov
-        uses: codecov/codecov-action@v3
+        uses: codecov/codecov-action@v4
         with:
           fail_ci_if_error: false
 
