### 2024-02-09 c91ba58
Apply suggestions from code review
Co-authored-by: Yaroslav Halchenko <debian@onerussian.com>
