+++ docs/reproin.rst
@@ -15,7 +15,7 @@ That will be a topic for another tutorial but meanwhile you can checkout `reproi
 Get Example Dataset
 -------------------
 
-This example uses a phantom dataset: `reproin_dicom.zip <https://osf.io/4jwk5/>`_ generated by the University of Arizona on their Siemens Skyra 3T with Syngo MR VE11c software on 2018_02_08.
+This example uses a phantom dataset: `reproin_dicom.zip <https://datasets.datalad.org/?dir=/repronim/heudiconv-reproin-example>`_ generated by the University of Arizona on their Siemens Skyra 3T with Syngo MR VE11c software on 2018_02_08.
 
 The ``REPROIN`` directory is a simple reproin-compliant DICOM (.dcm) dataset without sessions. 
 (Derived dwi images (ADC, FA etc.) that the scanner produced have been removed.::
