### 2024-02-13 4688912
Create convertall_custom.py as a demonstration for a derived heuristic with custom_seqinfo
