+++ heudiconv/heuristics/convertall.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import logging
 from typing import Optional
 
-from heudiconv.dicoms import dw
 from heudiconv.utils import SeqInfo
 
 lgr = logging.getLogger("heudiconv")
@@ -17,21 +16,6 @@ def create_key(
     if template is None or not template:
         raise ValueError("Template must be a valid format string")
     return (template, outtype, annotation_classes)
-
-
-def custom_seqinfo(wrapper: dw.Wrapper, series_files: list[str]) -> tuple[str, str]:
-    # Just a dummy demo for what custom_seqinfo could get/do
-    # for already loaded DICOM data, and including storing/returning
-    # the sample series file as was requested
-    # in https://github.com/nipy/heudiconv/pull/333
-    from nibabel.nicom.dicomwrappers import WrapperError
-
-    try:
-        affine = wrapper.affine.tobytes()
-    except WrapperError:
-        lgr.exception("Errored out while obtaining/converting affine")
-        affine = None
-    return affine, series_files[0]
 
 
 def infotodict(
