### 2024-02-15 9530dd7
Document how to release and add changelog entries
Fixes #732
