### 2024-02-15 709b104
Consistent casing for HeuDiConv and GitHub project names
