### 2024-02-23 5fb4888
Add basic statement for the state of the field to note that BIDS evolves and point to existing list of converters etc
