### 2024-02-24 ce76091
mention stats fig 3 while talking about state of the field
