### 2024-02-24 fd1f838
Adjust wording on heuristics page -- do not claim creating some skeleton
