+++ .github/workflows/test.yml
@@ -56,5 +56,6 @@ jobs:
         uses: codecov/codecov-action@v4
         with:
           fail_ci_if_error: false
+          token: ${{ secrets.CODECOV_TOKEN }}
 
 # vim:set et sts=2:
