+++ paper.bib
@@ -141,3 +141,22 @@
 	  doi = {10.48550/arXiv.2309.05768},
 	  url = {https://doi.org/10.48550/arXiv.2309.05768}
 }
+
+@article{ezbids,
+  title =		 {{ezBIDS}: {G}uided standardization of neuroimaging data
+                  interoperable with major data archives and
+                  platforms},
+  volume =		 {11},
+  ISSN =		 {2052-4463},
+  url =			 {http://dx.doi.org/10.1038/s41597-024-02959-0},
+  DOI =			 {10.1038/s41597-024-02959-0},
+  number =		 {1},
+  journal =		 {Scientific Data},
+  publisher =	 {Springer Science and Business Media LLC},
+  author =		 {Levitas, Daniel and Hayashi, Soichi and
+                  Vinci-Booher, Sophia and Heinsfeld, Anibal and
+                  Bhatia, Dheeraj and Lee, Nicholas and Galassi,
+                  Anthony and Niso, Guiomar and Pestilli, Franco},
+  year =		 {2024},
+  month =		 feb
+}
