### 2024-02-24 #5839 @yarikoptic
I think it is time now!  First of all thank you @marcelzwiers and @JLefortBesnard for the reviews and @britta-wstnr for keeping the finger on the pulse of this publication!
Since the time of the reviews, to address the concerns we have
- in the paper
  - decided to keep existing Figure 1 as it depicts "placement" of heudiconv in the workflow of a neuroimaging study
  - added Figure 2 to depict two possible "workflows" of using `heudiconv` - either semi-automated or fully automated in one shot, which we hope would clarify that core aspect of using `heudiconv`.
  - added section on the "Statement of the field" for basic pointers to other solutions available, and short description of its specific (full comparison would be too much for this intended to be short paper)
  - in `ReproIn` heuristic description added a one-liner  usage example which should not outdate itself too fast, and pointed to the Tutorials section in the documentation for more examples
  - extended Figure 3 with up to date usage/download stats to further show that HeuDiConv's usage is growing steadily
- in HeuDiConv
  - extended its [README.md](https://github.com/nipy/heudiconv?tab=readme-ov-file#howto-101) with figures from the paper to better guide visitors who encounter heudiconv from github
  - extended documentation with better [description of heuristic files](https://heudiconv.readthedocs.io/en/latest/heuristics.html)
  - adopted (with permission)  [Dianne Patterson’s University of Arizona tutorials](https://neuroimaging-core-docs.readthedocs.io/en/latest/pages/heudiconv.html#lesson-3-reproin-py)  into our documentation: https://heudiconv.readthedocs.io/en/latest/tutorials.html

I hope these changes address concerns raised by the review and overall improve heudiconv for its users.
