### 2024-02-24 bc2726f
Unify (lower) casing in 2 section names
