### 2024-02-26 f99594d
Show auto version -v (to be able to debug) and then take last line for the actual version boost
