### 2024-02-26 c41a49b
Fix the incorrect label on invocation of auto version
