+++ .github/workflows/release.yml
@@ -24,7 +24,7 @@ jobs:
           wget -O- https://github.com/intuit/auto/releases/download/v10.16.1/auto-linux.gz | gunzip > ~/auto
           chmod a+x ~/auto
 
-      - name: Check whether a release is due
+      - name: Query 'auto' on type of the release
         id: auto-version
         run: |
           # to be able to debug if something goes wrong
