### 2024-02-26 e9fe960
set -o pipefail + do not rely on bump being the last line -- parse from debug msg
