+++ .github/workflows/release.yml
@@ -28,8 +28,9 @@ jobs:
         id: auto-version
         run: |
           # to be able to debug if something goes wrong
+          set -o pipefail
           auto version -v | tee /tmp/auto-version
-          version="$(tail -n 1 /tmp/auto-version)"
+          version=$(sed -ne '/Calculated SEMVER bump:/s,.*: *,,p' < /tmp/auto-version)
           echo "version=$version" >> "$GITHUB_OUTPUT"
         env:
           GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
