### 2024-02-26 nipy/heudiconv#715 @yarikoptic
done now: that was done
