### 2024-02-26 d26e901
Be more verbose while invoking auto version
Co-authored-by: John T. Wodder II <jwodder@users.noreply.github.com>
