### 2024-02-26 17aad35
Do not bother with < + quote just in case
Co-authored-by: John T. Wodder II <jwodder@users.noreply.github.com>
