+++ .github/workflows/release.yml
@@ -30,7 +30,7 @@ jobs:
           # to be able to debug if something goes wrong
           set -o pipefail
           auto version -vv | tee /tmp/auto-version
-          version=$(sed -ne '/Calculated SEMVER bump:/s,.*: *,,p' < /tmp/auto-version)
+          version="$(sed -ne '/Calculated SEMVER bump:/s,.*: *,,p' /tmp/auto-version)"
           echo "version=$version" >> "$GITHUB_OUTPUT"
         env:
           GH_TOKEN: ${{ secrets.GITHUB_TOKEN }}
