### 2024-02-26 d6ed9b0
Fix - auto is in ~/, not in the PATH
