### 2024-02-28 41b22d7
Provide types for custom_seqinfo but disable type checking in the test -- I think mypy has deficiency since we do use kwargs and thus should be ok
