### 2024-03-05 0821456
try to fix code failure when dicom time fields are empty
This scenario happens when anonymization empties the field (not removed).
