### 2024-03-22 #5839 @britta-wstnr
Hey @JLefortBesnard ,
thanks so much for your review of the revision. For me, not all of the checkboxes in your [checklist](https://github.com/openjournals/joss-reviews/issues/5839#issuecomment-1779161965) appear checked off yet, can you have a look and confirm if those issues are resolved or not? Thanks a lot! 🙏
