### 2024-03-29 #5839 @britta-wstnr
A big thanks already to @marcelzwiers and @JLefortBesnard for their reviewing work! JOSS and I appreciate the time and work you put into this! 💙

@yarikoptic - can you please _double check authors and affiliations (including ORCIDs)_ on the paper? 🙏
The next step is for me to read the paper again. I might get back to you with some things (like typos) to fix.
After that, I will have some more things to do for you, so keep an eye out for notifications 🙂
