### 2024-03-29 #5839 @yarikoptic
I have clicked through all ORCIDs where provided -- seems to correspond.  Names are also fine but some might have preferences on either to use or not middle initial etc.  I think it is all good overall.  May be let's give a day for co-authors possibly to offer corrections.

Attn @mgxd,@satra,@pvelasco,@mvdoc,@tsalo,@jwodder,@mih,@psadil,@chrisgorgo,@TheChymera,@neurolabusc,@tjhendrickson,@mick-d,@jdkent,@tedstrauss,@leej3,@candleindark,@effigies,@darrencl,@butellyn,@mainux,@DVSneuro,@asmacdo,@dnkennedy

FTR list of coauthors produced using `jq -r 'map(select(.status == "co-author")) | .[].login' invitees.json | grep -v yarikoptic | while read l; do echo -n "@$l,"; done` after I updated the https://github.com/nipy/heudiconv/blob/paper-joss/authors/invitees.json (content is "private" in annex)
