### 2024-04-09 #5839 @yarikoptic
done in https://github.com/nipy/heudiconv/commit/b4184b15ee352c537c4945076e76f01659226aa2
