+++ paper.md
@@ -147,12 +147,12 @@ bibliography: paper.bib
 
 # Summary
 
-In order to support efficient processing, data must be formatted according to standards prevalent in the field, and widely supported among actively developed analysis tools.
+In order to support efficient processing, data must be formatted according to standards that are prevalent in the field and widely supported among actively developed analysis tools.
 The Brain Imaging Data Structure (BIDS) [@GAC+16] is an open standard designed for computational accessibility, operator legibility, and a wide and easily extendable scope of modalities — and is consequently used by numerous analysis and processing tools as the preferred input format in many fields of neuroscience.
 HeuDiConv (Heuristic DICOM Converter) enables flexible and efficient conversion of spatially reconstructed neuroimaging data from the DICOM format (quasi-ubiquitous in biomedical image acquisition systems, particularly in clinical settings) to BIDS, as well as other file layouts.
-HeuDiConv provides a multi-stage operator input workflow (discovery, manual tuning, conversion) where manual tuning step is optional and thus the entire conversion can be seamlessly integrated into a data processing pipeline.
+HeuDiConv provides a multi-stage operator input workflow (discovery, manual tuning, conversion) where a manual tuning step is optional and the entire conversion can thus be seamlessly integrated into a data processing pipeline.
 HeuDiConv is written in Python, and supports the DICOM specification for input parsing, and the BIDS specification for output construction.
-The support for these standards is extensive, and HeuDiConv can handle complex organization scenarios such as arise for specific data types (e.g., multi-echo sequences, or single-band reference volumes).
+The support for these standards is extensive, and HeuDiConv can handle complex organization scenarios that arise for specific data types (e.g., multi-echo sequences, or single-band reference volumes).
 In addition to generating valid BIDS outputs, additional support is offered for custom output layouts.
 This is obtained via a set of built-in fully functional or example heuristics expressed as simple Python functions.
 Those heuristics could be taken as a template or as a base for developing custom heuristics, thus providing full flexibility and maintaining user accessibility.
@@ -167,13 +167,13 @@ Neuroimaging data sourced from medical imaging equipment, and in particular magn
 DICOM data are often transmitted to PACS (Picture Archiving and Communication Systems) servers for archiving or further processing.
 Unlike in clinical settings, where data are interfaced with directly from PACS in the DICOM format, in neuroimaging research, tools typically require data files in the NIfTI [@nifticlib] format which directly stores images as 3D or 4D objects and restricts metadata to the most useful attributes.
 Tools such as `dcm2niix` [@Li_2016] can be used to convert DICOM files into NIfTI files, and can extract metadata fields not covered by the NIfTI header into sidecar `.json` files.
-However, the scope of such tools is limited, as it does not extend to organizing multiple NIfTI files for different subjects and possibly scanning sessions within a study.
+However, the scope of such tools is limited, as they do not extend to organizing multiple NIfTI files for different subjects and scanning sessions within a study.
 
 HeuDiConv was created in 2014 to provide flexible tooling so that labs may rapidly and consistently convert collections of DICOM files into collections of NIfTI files in customizable file system hierarchies.
 As manual file renaming and metadata reorganization is tedious and error prone, automation is preferable, and this is a consistent focus of HeuDiConv.
 
 Since the inception of HeuDiConv in 2014, the BIDS standard [@GAC+16] was established.
[…]
