### 2024-04-16 49c5783
sequence_name from PulseSequenceName on Siemens XA series, need tests data to write tests
