### 2024-04-16 9afeae9
sequence_name from PulseSequenceName on Siemens XA series, need tests data to write tests
