### 2024-04-24 nipy/heudiconv#754 @yarikoptic
ORCID for  JOSS publication for @toddt
Dear @toddt

Per your response at https://github.com/nipy/heudiconv-joss-paper/issues/44#issuecomment-1599488964 we did add you ["in the middle somewhere"](https://github.com/nipy/heudiconv/blob/paper-joss/paper.md?plain=1#L78).

But now the sharp eye of our tireless editor @britta-wstnr [has mentioned](https://github.com/nipy/heudiconv/pull/750#issue-2235192432) lack of ORCID entry for you.  Do you have ORCID record to be used in the publication?

Thank you in advance for the rapid communication
