### 2024-04-24 4583f60
Add missing middle initial J for ChrisG in one publication
