+++ paper.bib
@@ -116,7 +116,7 @@
 	year = 2011,
 	publisher = {Frontiers Media {SA}},
 	volume = {5},
-	author = {Krzysztof Gorgolewski and Christopher D. Burns and Cindee Madison and Dav Clark and Yaroslav O. Halchenko and Michael L. Waskom and Satrajit S. Ghosh},
+	author = {Krzysztof J. Gorgolewski and Christopher D. Burns and Cindee Madison and Dav Clark and Yaroslav O. Halchenko and Michael L. Waskom and Satrajit S. Ghosh},
 	title = {Nipype: A Flexible, Lightweight and Extensible Neuroimaging Data Processing Framework in Python},
 	journal = {Frontiers in Neuroinformatics}
 }
