+++ paper.bib
@@ -131,15 +131,68 @@
   copyright = {Open Access}
 }
 
-@article{PMA+arXiv2024:bids-past,
-      title={The Past, Present, and Future of the Brain Imaging Data Structure (BIDS)},
-      author={Russell A. Poldrack and Christopher J. Markiewicz and Stefan Appelhoff and Yoni K. Ashar and Tibor Auer and Sylvain Baillet and Shashank Bansal and Leandro Beltrachini and Christian G. Benar and Giacomo Bertazzoli and Suyash Bhogawar and Ross W. Blair and Marta Bortoletto and Mathieu Boudreau and Teon L. Brooks and Vince D. Calhoun and Filippo Maria Castelli and Patricia Clement and Alexander L Cohen and Julien Cohen-Adad and Sasha D'Ambrosio and Gilles de Hollander and María de la iglesia-Vayá and Alejandro de la Vega and Arnaud Delorme and Orrin Devinsky and Dejan Draschkow and Eugene Paul Duff and Elizabeth DuPre and Eric Earl and Oscar Esteban and Franklin W. Feingold and Guillaume Flandin and anthony galassi and Giuseppe Gallitto and Melanie Ganz and Rémi Gau and James Gholam and Satrajit S. Ghosh and Alessio Giacomel and Ashley G Gillman and Padraig Gleeson and Alexandre Gramfort and Samuel Guay and Giacomo Guidali and Yaroslav O. Halchenko and Daniel A. Handwerker and Nell Hardcastle and Peer Herholz and Dora Hermes and Christopher J. Honey and Robert B. Innis and Horea-Ioan Ioanas and Andrew Jahn and Agah Karakuzu and David B. Keator and Gregory Kiar and Balint Kincses and Angela R. Laird and Jonathan C. Lau and Alberto Lazari and Jon Haitz Legarreta and Adam Li and Xiangrui Li and Bradley C. Love and Hanzhang Lu and Camille Maumet and Giacomo Mazzamuto and Steven L. Meisler and Mark Mikkelsen and Henk Mutsaerts and Thomas E. Nichols and Aki Nikolaidis and Gustav Nilsonne and Guiomar Niso and Martin Norgaard and Thomas W Okell and Robert Oostenveld and Eduard Ort and Patrick J. Park and Mateusz Pawlik and Cyril R. Pernet and Franco Pestilli and Jan Petr and Christophe Phillips and Jean-Baptiste Poline and Luca Pollonini and Pradeep Reddy Raamana and Petra Ritter and Gaia Rizzo and Kay A. Robbins and Alexander P. Rockhill and Christine Rogers and Ariel Rokem and Chris Rorden and Alexandre Routier and Jose Manuel Saborit-Torres and Taylor Salo and Michael Schirner and Robert E. Smith and Tamas Spisak and Julia Sprenger and Nicole C. Swann and Martin Szinte and Sylvain Takerkart and Bertrand Thirion and Adam G. Thomas and Sajjad Torabian and Gael Varoquaux and Bradley Voytek and Julius Welzel and Martin Wilson and Tal Yarkoni and Krzysztof J. Gorgolewski},
-      year={2024},
-      eprint={2309.05768},
-      archivePrefix={arXiv},
-      primaryClass={q-bio.OT},
-	  doi = {10.48550/arXiv.2309.05768},
-	  url = {https://doi.org/10.48550/arXiv.2309.05768}
+@article{PMA+24:bids-past,
+  title =		 {The past, present, and future of the brain imaging
+                  data structure (BIDS)},
+  volume =		 {2},
+  ISSN =		 {2837-6056},
[…]
