### 2024-04-24 26072be
Split a sentence and make another one "complete"
