### 2024-04-24 #5839 @yarikoptic
done -- see https://github.com/nipy/heudiconv/pull/750#issuecomment-2075686012 , let's finalize that one and wait a day or so for
- https://github.com/nipy/heudiconv/issues/754
