### 2024-04-25 68b4e31
Adjust wording a little in Figure caption to make easier to read (from @britta-wstnr)
