+++ paper.md
@@ -193,7 +193,7 @@ Such decision making is implemented in *HeuDiConv heuristics*, which are impleme
 HeuDiConv, if instructed to operate in BIDS mode (`--bids` flag) with a heuristic providing base naming instructions, helps to organize the files in the hierarchy defined by the BIDS standard.
 It also ensures files are named according to the BIDS specifications, including complex composite recordings such as those associated with multi-echo sequences.
 
-![**HeuDiConv automates the keystone conversion step in reproducible data handling, without compromising operator flexibility.** The showcased set-up depicts a 2-machine infrastructure, with heudiconv operating on the same machine as subsequent analysis steps for data in a standardized and shareable representation. For more advanced usage at institutions with dedicated infrastructure, HeuDiConv can operate on an additional third machine, interfacing between the depicted two, and dedicated to data repositing, versioning, and backup.](figs/environment.pdf)
+![**HeuDiConv automates the keystone conversion step in reproducible data handling, without compromising operator flexibility.** The showcased set-up depicts a 2-machine infrastructure, with heudiconv operating on the same machine as subsequent analysis steps for data in a standardized and shareable representation. For more advanced usage at institutions with dedicated infrastructure, HeuDiConv can operate on an additional third machine, which then interfaces between the depicted two machines and is dedicated to data repositing, versioning, and backup.](figs/environment.pdf)
 
 
 ![**HeuDiConv conversion and layout is controlled via heuristics (custom or provided built-ins) either with manual tune up of proposed filenames or fully automated.** The heudiconv application can be used with the `-c none` parameter to generate by heuristic a list of filenames for the user to edit, before invoking the conversion to be performed via the `-c dcm2niix` option to use the `dcm2niix` tool. The process is idempotent, and specifying the `-c dcm2niix` option can automatically convert without seeking user tune up of proposed filenames.](figs/workflow.pdf)
