### 2024-04-30 nipy/heudiconv#754 @yarikoptic
let's close this window of the opportunity
