### 2024-05-01 #5839 @britta-wstnr
Hi @yarikoptic, thanks for the paper fixes!
Let's move on :)
Could you take care of the **Additional Author Tasks** in [this checklist](https://github.com/openjournals/joss-reviews/issues/5839#issuecomment-2027036743)? The first one is already done, so it's now all about the archiving and versions.
Once that's done, please give me a ping here - then I will go on with the last few tasks!
