### 2024-05-01 b3cfd73
RF: replace string size comparison with a more generic "bool" of a string check
