### 2024-05-01 c29047a
Update heudiconv/dicoms.py
Co-authored-by: Yaroslav Halchenko <debian@onerussian.com>
