### 2024-05-01 f3a6d5f
Update heudiconv/dicoms.py
Co-authored-by: Yaroslav Halchenko <debian@onerussian.com>
