### 2024-05-01 7342206
Update heudiconv/dicoms.py
Co-authored-by: Yaroslav Halchenko <debian@onerussian.com>
