### 2024-05-06 #5839 @britta-wstnr
Hi @yarikoptic, I am not aware of any package or software that does this automagically. But maybe [this script from MNE-Python](https://github.com/mne-tools/mne-python/blob/main/tools/generate_codemeta.py) can server as inspiration? This one works on the git history though.
