### 2024-05-06 #5839 @oliviaguest
> > Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
>
> oopsy... are you @britta-wstnr aware of a helper to produce a CITATION.cff (may be @jspaaks knows) or `.zenodo.json` (may be @lnielsen knows) out of JOSS'es paper header?  Apparently we lack any of those for heudiconv so we are out of sync here.
>
> edit1: oh -- 3 years ago I created
>
> - https://github.com/con/tributors/issues/57
>
> eh heh... might be time is ripe
>
> > Make sure that the license listed for the archive is the same as the software license.
>
> and there too :-/

Hi @yarikoptic I'm just jumping in here so I may be wrong or just repeating what @britta-wstnr said from a different angle, but I suspect you can just wait for the bot to make one for you? For example: https://github.com/openjournals/joss-reviews/issues/6186#issuecomment-2094758078
