### 2024-05-06 #5839 @yarikoptic
> Hi @yarikoptic I'm just jumping in here so I may be wrong or just repeating what @britta-wstnr said from a different angle, but I suspect you can just wait for the bot to make one for you? For example: [#6186 (comment)](https://github.com/openjournals/joss-reviews/issues/6186#issuecomment-2094758078)

thank you @oliviaguest !  It seems though to be a chicken&egg issue somewhat now - I need zenodo entry to align before this is accepted! ;) good to know that then I would get CITATION.cff I could add/reuse... might want to look into bot code to just trigger it locally.

Anyone knows the location of the bot code? (hard to find it by searching for produced content since it is all over the github ;))

> Hi @yarikoptic, I am not aware of any package or software that does this automagically. But maybe [this script from MNE-Python](https://github.com/mne-tools/mne-python/blob/main/tools/generate_codemeta.py) can server as inspiration? This one works on the git history though.

Thanks for the reference! indeed a bit too custom/ad-hoc but could give ideas. Thanks!
