### 2024-05-07 #5839 @oliviaguest
> Anyone knows the location of the bot code? (hard to find it by searching for produced content since it is all over the github ;))

[My strong suggestion is, along the lines of the reply by @britta-wstnr, to focus on this matter right now (i.e. whatever is left for your JOSS paper to be published), but since you asked: https://github.com/openjournals/buffy.]
