### 2024-05-07 #5839 @jspaaks
> oopsy... are you @britta-wstnr aware of a helper to produce a CITATION.cff (may be @jspaaks knows) or `.zenodo.json` (may be @lnielsen knows) out of JOSS'es paper header? Apparently we lack any of those for heudiconv so we are out of sync here.

Sorry @yarikoptic as far as I'm aware that doesn't exist
