### 2024-05-15 76e731d
[DATALAD RUNCMD] Produce CITATION.cff from JOSS paper
=== Do not change lines below ===
{
 "chain": [],
 "cmd": "../heudiconv-joss-paper/tools/paper-to-citationcff ../heudiconv-joss-paper/paper.md",
 "exit": 0,
 "extra_inputs": [],
 "inputs": [],
 "outputs": [],
 "pwd": "."
}
^^^ Do not change lines above ^^^
