+++ CITATION.cff
@@ -1,0 +1,141 @@
+authors:
+- affiliation: Center for Open Neuroscience, Department of Psychological and Brain
+    Sciences, Dartmouth College, Hanover, NH, USA
+  family-names: Halchenko
+  given-names: Yaroslav O.
+  orcid: https://orcid.org/0000-0003-3456-2493
+- affiliation: Department of Psychology, Stanford University, CA, USA
+  family-names: Goncalves
+  given-names: Mathias
+  orcid: https://orcid.org/0000-0002-7252-7771
+- affiliation: McGovern Institute, Massachusetts Institute of Technology, Cambridge,
+    MA, USA
+  family-names: Ghosh
+  given-names: Satrajit
+  orcid: https://orcid.org/0000-0002-5312-6729
+- affiliation: Flywheel Exchange LLC, Minneapolis, MN, USA
+  family-names: Velasco
+  given-names: Pablo
+  orcid: https://orcid.org/0000-0002-5749-6049
+- affiliation: University of California, Berkeley, Berkeley, CA, USA
+  family-names: Castello
+  given-names: Matteo Visconti di Oleggio
+  orcid: https://orcid.org/0000-0001-7931-5272
+- affiliation: Perelman School of Medicine, University of Pennsylvania, Philadelphia,
+    PA, USA
+  family-names: Salo
+  given-names: Taylor
+  orcid: https://orcid.org/0000-0001-9813-3167
+- affiliation: Center for Open Neuroscience, Department of Psychological and Brain
+    Sciences, Dartmouth College, Hanover, NH, USA
+  family-names: II
+  given-names: John T. Wodder
+- affiliation: "Institute of Neuroscience and Medicine, Brain & Behaviour (INM-7),\
+    \ Research Center J\xFClich, J\xFClich, Germany & Institute of Systems Neuroscience,\
+    \ Medical Faculty, Heinrich Heine University D\xFCsseldorf, D\xFCsseldorf, Germany"
+  family-names: Hanke
+  given-names: Michael
+  orcid: https://orcid.org/0000-0001-6398-6370
+- affiliation: Department of Biostatistics, Johns Hopkins Bloomberg School of Public
+    Health, Baltimore, MD, USA
+  family-names: Sadil
+  given-names: Patrick
+  orcid: https://orcid.org/0000-0003-4141-1343
+- affiliation: Emeritus of Department of Psychology, Stanford University, CA, USA
+  family-names: Gorgolewski
+  given-names: Krzysztof Jacek
+  orcid: https://orcid.org/0000-0003-3321-7583
+- affiliation: Center for Open Neuroscience, Department of Psychological and Brain
+    Sciences, Dartmouth College, Hanover, NH, USA
+  family-names: Ioanas
+  given-names: Horea-Ioan
+  orcid: https://orcid.org/0000-0001-7037-2449
+- affiliation: Department of Psychology, University of South Carolina, Columbia, SC,
+    USA
+  family-names: Rorden
+  given-names: Chris
+  orcid: https://orcid.org/0000-0002-7554-6142
+- affiliation: "Masonic Institute\_for the Developing Brain, University of Minnesota,\
+    \ Minneapolis, MN, USA & Minnesota Supercomputing Institute, University of Minnesota,\
+    \ Minneapolis, MN, USA"
+  family-names: Hendrickson
+  given-names: Timothy J.
+  orcid: https://orcid.org/0000-0001-6862-6526
+- affiliation: Human Neuroscience Platform, Fondation Campus Biotech Geneva, Geneva,
+    Switzerland
+  family-names: Dayan
+  given-names: Michael
+  orcid: https://orcid.org/0000-0002-2666-0969
+- affiliation: Center for Open Neuroscience, Department of Psychological and Brain
[…]
