+++ CITATION.cff
@@ -18,8 +18,8 @@ authors:
   given-names: Pablo
   orcid: https://orcid.org/0000-0002-5749-6049
 - affiliation: University of California, Berkeley, Berkeley, CA, USA
-  family-names: Castello
-  given-names: Matteo Visconti di Oleggio
+  family-names: Visconti di Oleggio Castello
+  given-names: Matteo
   orcid: https://orcid.org/0000-0001-7931-5272
 - affiliation: Perelman School of Medicine, University of Pennsylvania, Philadelphia,
     PA, USA
@@ -28,8 +28,8 @@ authors:
   orcid: https://orcid.org/0000-0001-9813-3167
 - affiliation: Center for Open Neuroscience, Department of Psychological and Brain
     Sciences, Dartmouth College, Hanover, NH, USA
-  family-names: II
-  given-names: John T. Wodder
+  family-names: Wodder II
+  given-names: John T.
 - affiliation: "Institute of Neuroscience and Medicine, Brain & Behaviour (INM-7),\
     \ Research Center J\xFClich, J\xFClich, Germany & Institute of Systems Neuroscience,\
     \ Medical Faculty, Heinrich Heine University D\xFCsseldorf, D\xFCsseldorf, Germany"
