### 2024-05-16 #5839 @yarikoptic
> > Anyone knows the location of the bot code? (hard to find it by searching for produced content since it is all over the github ;))
>
> [My strong suggestion is, along the lines of the reply by @britta-wstnr, to focus on this matter right now (i.e. whatever is left for your JOSS paper to be published), but since you asked: https://github.com/openjournals/buffy.]

The matter at hands remains the same (and some aspects of that we are addressing with heudiconv) -- **humans suck** with entering or even copy/pasting metadata...  So I ended up quickly chatting with chatgpt and coming up with https://github.com/nipy/heudiconv/blob/paper-joss/tools/paper-to-citationcff which I used to produce https://github.com/nipy/heudiconv/blob/master/CITATION.cff which I tuned up just a bit.  Now someone could generalize it more or just use as is.

With that let's get back to our checklist

#### Additional Author Tasks After Review is Complete
- [x] Make a release of the software with the latest changes from the review and post the version number here: 1.1.3
- [x] Archive the release on Zenodo/figshare/etc and post the DOI here: https://doi.org/10.5281/zenodo.11201247
- [x] Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper. (**note: zenodo uses `Last, First` whenever JOSS `First Last` so it matches, given correct reassembly)
- [x] Make sure that the license listed for the archive is the same as the software license. (required 1.1.3 to chase 1.1.2 ;-) )
