### 2024-05-21 #5839 @britta-wstnr
Hi @yarikoptic - using my non-AI powers to double check the archive and paper authors lists 😉 I noticed the following:
- It seems like your AI friend omitted any double affiliations from your archive list and only gives the first.
- In your paper, the affiliations are not ordered (probably based on how you define the author order and affiliation list in your `paper.md`. I strongly suggest ordering them (as Zenodo automatically does) for a clean look.

Thanks! 🙏
