### 2024-05-21 #5839 @yarikoptic
> * It seems like your AI friend omitted any double affiliations from your archive list and only gives the first.

The problem is that CITATION.cff does not allow for multiple affiliations

- https://github.com/citation-file-format/citation-file-format/issues/268

so I had to join them with a custom `&` into a single one at https://github.com/nipy/heudiconv/blob/paper-joss/tools/paper-to-citationcff#L46 , thus in those cases generating a new unique meta-affiliation.  I hope that is ok

> * In your paper, the affiliations are not ordered (probably based on how you define the author order and affiliation list in your `paper.md`. I strongly suggest ordering them (as Zenodo automatically does) for a clean look.

Sorry, I am not entirely following -- ordered how exactly?  indeed I define them in order of encountering through the author list, hence

![image](https://github.com/openjournals/joss-reviews/assets/39889/1aa15575-a532-4273-9c3c-d17fb5d6c6f4)

and I think zenodo does the same:

![image](https://github.com/openjournals/joss-reviews/assets/39889/71dea3ad-353a-45ec-86a3-213db432f99c)

If there is no specific requirement by JOSS, I would prefer to keep this as is.
