### 2024-05-21 #5839 @britta-wstnr
@yarikoptic if you look further down the list, your JOSS affiliations jump in order:
![image](https://github.com/openjournals/joss-reviews/assets/20129119/5dcd288e-2eaa-4768-8670-bd192cd0553f)
7, 8, 22, 24, 9 ...

The Zenodo ones do not do that.
As for the merged affiliations on Zenodo - I am not sure, @oliviaguest what do you think?
