### 2024-05-21 #5839 @yarikoptic
oh, a sharp eye @britta-wstnr !  Indeed we have it in that non-harmoniously increasing order in the
https://github.com/nipy/heudiconv/blob/paper-joss/paper.md?plain=1#L33 . Given the pain of redoing all of that

- if there is no requirement for their ordering: let's proceed as is;
- if there is a requirement -- we better also file an issue that bot/tools either order automagically or at least warn/error if order is not non-harmoniously increasing.

zenodo has no issue because [CITATION.cff](https://github.com/nipy/heudiconv/blob/master/CITATION.cff) doesn't have any numbering and rather explicit statement of the affiliation and then does deduplication and indexing "internally".
