### 2024-05-30 e029849
add complex part to ImageType for use in heuristics, it is done in dcm2niix when converting
