### 2024-05-30 3441c53
add complex part to ImageType for use in heuristics, it is done in dcm2niix when converting
