+++ heudiconv/dicoms.py
@@ -111,6 +111,15 @@ def create_seqinfo(
     else:
         sequence_name = ""
 
+    # GE data
+    # see https://github.com/rordenlab/dcm2niix/tree/master/GE#complex-image-component
+    if dcminfo.get([0x43, 0x102F]):
+        GE_CPLX_CODING = ["PHASE", "MAGNITUDE", "REAL", "IMAGINARY"]
+        cplx_idx = int(dcminfo.get([0x43, 0x102F]).value)
+        part = GE_CPLX_CODING[cplx_idx]
+        if part not in image_type:
+            image_type = image_type + (part,)
+
     # initialized in `group_dicoms_to_seqinfos`
     global total_files
     total_files += len(series_files)
