### 2024-05-30 9a51164
more fixes, works on example GE data, require dcm2niix patch
