### 2024-05-30 39dfd5b
more fixes, works on example GE data, require dcm2niix patch
