### 2024-05-30 7f82416
more fixes, works on example GE data, require dcm2niix patch
