### 2024-05-30 8a5d27c
more fixes, works on example GE data, require dcm2niix patch
