### 2024-05-30 80e62bc
Fix up typos detected with fresh codespell
Thanks @bpinsard for the buzz
