### 2024-05-31 9f2cf4b
Update heudiconv/convert.py
Co-authored-by: Yaroslav Halchenko <debian@onerussian.com>
