### 2024-06-05 3678f3e
[DATALAD RUNCMD] Fixup affiliations
=== Do not change lines below ===
{
 "chain": [],
 "cmd": "tools/joss-paper-buddy fixup-affiliations",
 "dsid": "c8375293-3dcf-49d6-a146-563ac382ff71",
 "exit": 0,
 "extra_inputs": [],
 "inputs": [],
 "outputs": [],
 "pwd": "."
}
^^^ Do not change lines above ^^^
