+++ paper.md
@@ -1,13 +1,13 @@
 ---
-title: 'HeuDiConv — flexible DICOM conversion into structured directory layouts'
+title: HeuDiConv — flexible DICOM conversion into structured directory layouts
 tags:
-  - Python
-  - neuroscience
-  - standardization
-  - DICOM
-  - BIDS
-  - open science
-  - FOSS
+ - Python
+ - neuroscience
+ - standardization
+ - DICOM
+ - BIDS
+ - open science
+ - FOSS
 authors:
  - name: Yaroslav O. Halchenko  # 626 commits, issues: opened 81 participated in 146
    orcid: 0000-0003-3456-2493
@@ -28,41 +28,41 @@ authors:
    orcid: 0000-0001-9813-3167
    affiliation: 6
  - name: John T. Wodder II  # 58 commits, issues: opened 2 participated in 4
-   orcid: 
+   orcid:
    affiliation: 1
  - name: Michael Hanke  # 32 commits, issues: opened 10 participated in 6
    orcid: 0000-0001-6398-6370
    affiliation: 7, 8
  - name: Patrick Sadil  # 18 commits, issues: opened 3 participated in 2
    orcid: 0000-0003-4141-1343
-   affiliation: 22
+   affiliation: 9
  - name: Krzysztof Jacek Gorgolewski  # 4 commits, issues: opened 7 participated in 10
    orcid: 0000-0003-3321-7583
-   affiliation: 24
+   affiliation: 10
  - name: Horea-Ioan Ioanas   # 17 commits, issues: opened 1 participated in 1
    orcid: 0000-0001-7037-2449
    affiliation: 1
  - name: Chris Rorden  # 0 commits, issues: opened 1 participated in 17
    orcid: 0000-0002-7554-6142
-   affiliation: 9
+   affiliation: 11
  - name: Timothy J. Hendrickson  # 0 commits, issues: opened 10 participated in 4
    orcid: 0000-0001-6862-6526
-   affiliation: 10, 11
+   affiliation: 12, 13
  - name: Michael Dayan  # 12 commits, issues: opened 1 participated in 2
    orcid: 0000-0002-2666-0969
-   affiliation: 12
+   affiliation: 14
  - name: Sean Dae Houlihan  # 12 commits, issues: opened 0 participated in 0 # First: Sean Dae; Last: Houlihan
    orcid: 0000-0001-5003-9278
-   affiliation:  1, 13
+   affiliation: 1, 15
  - name: James Kent  # 8 commits, issues: opened 3 participated in 0
    orcid: 0000-0002-4892-2659
-   affiliation: 14
+   affiliation: 16
  - name: Ted Strauss  # 0 commits, issues: opened 8 participated in 0
    orcid: 0000-0002-1927-666X
-   affiliation: 15
+   affiliation: 17
  - name: John Lee  # 2 commits, issues: opened 5 participated in 1
    orcid: 0000-0001-5884-4247
-   affiliation: 16
+   affiliation: 18
  - name: Isaac To  # 7 commits, issues: opened 0 participated in 0
    orcid: 0000-0002-4740-0824
    affiliation: 1
@@ -71,25 +71,25 @@ authors:
    affiliation: 2
  - name: Darren Lukas  # 4 commits, issues: opened 1 participated in 2
    orcid: 0009-0003-6941-0833
-   affiliation: 17
+   affiliation: 19
  - name: Ellyn R. Butler  # 0 commits, issues: opened 4 participated in 2
    orcid: 0000-0001-6316-6444
-   affiliation: 23
+   affiliation: 20
  - name: Todd Thompson  # 0 commits, issues: opened 5 participated in 0
-   orcid: 
[…]
