### 2024-06-05 fcce672
Remove CITATION.cff to please zenodo to please JOSS
References:
- https://github.com/openjournals/joss-reviews/issues/5839
- https://github.com/citation-file-format/citation-file-format/issues/374
