### 2024-06-05 #5839 @yarikoptic
- I have now pushed updated/renamed version of the helper now called [joss-paper-buddy](https://github.com/nipy/heudiconv/blob/paper-joss/tools/joss-paper-buddy) which could be used by others to
  - reorder affiliations -- done in https://github.com/nipy/heudiconv/commit/3678f3e1949682a7b090e3718cd30186089d3425
  - render CITATION.cff
  - render .zenodo.json -- done in https://github.com/nipy/heudiconv/pull/764 which also triggered a release of 1.1.4
- pipy got 1.1.4: https://pypi.org/project/heudiconv/
- but zenodo didn't .... so I guess something in .zenodo.json tripped it... nope -- somehow it was disabled
![image](https://github.com/openjournals/joss-reviews/assets/39889/083c02a9-d04e-49f7-97bc-bbfeafe7b639)
**fun!**   will look into it more later ... bleh....
