### 2024-06-06 256495f
Revert "Add CITATION.cff validation to lint workflow"
This reverts commit dcff5666bd789f7b1eff1ec33a4952fdb87aa389.

As we no longer have CITATION.cff and this was causing linters to fail.

We also need to trigger yet another release since zenodo integration for some
reason was off.
