+++ .github/workflows/lint.yml
@@ -22,10 +22,6 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           python -m pip install --upgrade tox
-          python -m pip install --upgrade cffconvert pyyaml
 
       - name: Run linters
         run: tox -e lint
-
-      - name: Validate CITATION.cff
-        run: cffconvert --validate
