### 2024-06-06 #5839 @yarikoptic
After all the dances with a drum I bring you 1.1.6 with

![image](https://github.com/openjournals/joss-reviews/assets/39889/e817c1ee-b496-4aa7-b934-54f9d6765e14)

which still does not match the paper!

![image](https://github.com/openjournals/joss-reviews/assets/39889/16c05a23-654c-4780-8d2a-50aa66442257)

and just now realized that they remained unmerged. But Dear @britta-wstnr and @crvernon -- are you sure it is technically possible at all to achieve unmerged multiple affiliations on zenodo?  Looking at https://developers.zenodo.org/#representation I see only a single affiliation string to be provided for each creator

![image](https://github.com/openjournals/joss-reviews/assets/39889/fbe061dc-fcc3-466b-a85f-973f7dc76652)

so seemingly not only CITATION.cff which I had to ditch to satisfy the target demand, nor seems .zenodo.json support that.  Please advise!

<summary>Precedence: in datalad JOSS paper we had multiple affiliations in JOSS paper, not even in the same order...</summary>

![image](https://github.com/openjournals/joss-reviews/assets/39889/db33a4af-7722-49b2-a163-b3fe3251e172)

and joined ones on zenodo "just fine"

![image](https://github.com/openjournals/joss-reviews/assets/39889/3592256b-ea46-43e5-8a2c-13be224e675a)
