### 2024-06-10 #5839 @yarikoptic
HA! that is a nice find @britta-wstnr ! Nothing is impossible -- just some times requires magic (or technology; as I hinted above on needing it a the level of the framework) ;)

Given that their [CITATION.cff](https://github.com/openneuropet/PET2BIDS/blob/main/CITATION.cff) has only

```yaml
- family-names: "Galassi"
  given-names: "Anthony"
  orcid: "https://orcid.org/0000-0001-6550-4574"
```

ORCID entries without affiliations, I guess zenodo is fetching from ORCID!  @CPernet and @bendhouseart -- how did you dare to be so smart and not to disclose this nice solution during our last beertogether?! ;-)

As I hate duplication (and here it is kinda the cause of the issue), it is a very interesting idea to pursue -- for authors to simply provide ORCIDs and completely avoid specification of affiliations in the paper and their software records, so that systems pick  up up-to-date affiliations from ORCID and thus completely avoid chances for divergence, typos, etc.
But for that more centralized effort is needed, hence ping @crvernon !

Getting to our little publication at hands, I will try **but** I am afraid it might be disastrous given the number of co-authors and likely not all of them having up-to-date ORCID entries etc... will report back on that.
