### 2024-06-10 #5839 @yarikoptic
ok, I have tried on https://github.com/yarikoptic/test/ with a copy of CITATION.cff I prepared for heudiconv, but where I removed all affiliations, so I got similar

```
- family-names: Halchenko
  given-names: Yaroslav O.
  orcid: https://orcid.org/0000-0003-3456-2493
```

and I got myself zenodo record: https://zenodo.org/records/11551364 but it lacks affiliations!

![image](https://github.com/openjournals/joss-reviews/assets/39889/f347d046-e5b7-4026-96d8-2d6a4e181f71)

I even then took PET2BIDS CITATION.cff and published that on zenodo to get https://zenodo.org/records/11551454 and also got no affiliations there.

so there is more of magic to it and CITATION.cff is not sufficient. Only @CPernet or @bendhouseart could help us out ATM I think ;)

NB I feel we are so far away from fulfilling the requirements for JOSS here...
