### 2024-06-11 #5839 @britta-wstnr
:wave: @yarikoptic - since our goal is to make sure we get you to publication status successfully, let's move forward with this by simply having you add the multiple affiliations per author manually in Zenodo.  This is very easy to do and would save us a lot of time in this case. Let me know if you need more guidance in how to do that.

I appreciate that this is cumbersome.  However, in this case, we should really get this one moving forward and dedicated a few minutes to do this should clear up the issue.

Thanks a lot - let me know if you need more information on this!
