### 2024-06-11 #5839 @yarikoptic
> by simply having you add the multiple affiliations per author manually in Zenodo. This is very easy to do and would save us a lot of time in this case. Let me know if you need more guidance in how to do that.

yes, just point me in the direction on how to do it manually please since I do not see any "Edit" button on https://zenodo.org/records/11497270 .
