### 2024-06-11 08c144f
[ENH] add PlainAcquisitionLabel IntendedFor method
Addresses #767

Do not use mixed acq/task entity for matching method as implemented for CustomAcquisitionLabel
