### 2024-06-12 #5839 @britta-wstnr
Hi @yarikoptic,
to be able to edit, you need to be the person with the log in credentials. See here for some more info: https://help.zenodo.org/docs/deposit/manage-records/

Once you got to the edit stage, you should be able to add several affiliations:

Hope this helps - let me know if you have any more questions!
