### 2024-06-14 #5839 @bendhouseart
> how did you dare to be so smart and not to disclose this nice solution during our last beertogether?! ;-)

I'm flattered, but the solution we used was not so smart as I manually entered everything on Zenodo. It's possible to add a zenodo.yaml (or whatever extension it is) and set things up such that each new tag/release will be automatically uploaded to zenodo with authors and all their affiliations. One _should_ probably do that, but I guess it depends on 1) how many extra config files one wants to have within their project 2) whether it's a repeat task worth automating.
