### 2024-06-20 #5839 @yarikoptic
long story short -- finally got permissions and manually edited that heudiconv zenodo entry -- should match now:

![image](https://github.com/openjournals/joss-reviews/assets/39889/283168b4-5ba2-4c58-a037-463cf550cb31)
