+++ heudiconv/heuristics/reproin.py
@@ -403,64 +403,61 @@ def infotodict(
     run_label: Optional[str] = None  # run-
     dcm_image_iod_spec: Optional[str] = None
     skip_derived = False
-    for i_acq, s in enumerate(seqinfo):
+    for i_acq, curr_seqinfo in enumerate(seqinfo):
         # XXX: skip derived sequences, we don't store them to avoid polluting
         # the directory, unless it is the motion corrected ones
         # (will get _rec-moco suffix)
-        if skip_derived and s.is_derived and not s.is_motion_corrected:
-            skipped.append(s.series_id)
-            lgr.debug("Ignoring derived data %s", s.series_id)
+        if (
+            skip_derived
+            and curr_seqinfo.is_derived
+            and not curr_seqinfo.is_motion_corrected
+        ):
+            skipped.append(curr_seqinfo.series_id)
+            lgr.debug("Ignoring derived data %s", curr_seqinfo.series_id)
             continue
 
         if i_acq == 0:
+            prev_seqinfo = None
             prev_dcm_image_iod_spec = None
-            prev_image_type_datatype = None
         else:
-            prev_dcm_image_iod_spec = seqinfo[i_acq - 1].image_type[2]
-            prev_image_type_datatype = {
-                # Note: P and M are too generic to make a decision here, could be
-                #  for different datatypes (bold, fmap, etc)
-                "FMRI": "func",
-                "MPR": "anat",
-                "DIFFUSION": "dwi",
-                "MIP_SAG": "anat",  # angiography
-                "MIP_COR": "anat",  # angiography
-                "MIP_TRA": "anat",  # angiography
-            }.get(prev_dcm_image_iod_spec, None)
+            prev_seqinfo = seqinfo[i_acq - 1]
+            for f in "series_description", "protocol_name":
+                prev_seqinfo = prev_seqinfo._replace(
+                    **{f: getattr(prev_seqinfo, f).format(**prev_seqinfo._asdict())}
+                )
+
+            prev_dcm_image_iod_spec = prev_seqinfo.image_type[2]
 
         if i_acq == (len(seqinfo) - 1):
+            next_seqinfo = None
             next_dcm_image_iod_spec = None
-            next_image_type_datatype = None
         else:
-            next_dcm_image_iod_spec = seqinfo[i_acq + 1].image_type[2]
-            next_image_type_datatype = {
-                # Note: P and M are too generic to make a decision here, could be
-                #  for different datatypes (bold, fmap, etc)
-                "FMRI": "func",
-                "MPR": "anat",
-                "DIFFUSION": "dwi",
-                "MIP_SAG": "anat",  # angiography
-                "MIP_COR": "anat",  # angiography
-                "MIP_TRA": "anat",  # angiography
-            }.get(next_dcm_image_iod_spec, None)
+            next_seqinfo = seqinfo[i_acq + 1]
+            for f in "series_description", "protocol_name":
+                next_seqinfo = next_seqinfo._replace(
+                    **{f: getattr(next_seqinfo, f).format(**next_seqinfo._asdict())}
+                )
+
+            next_dcm_image_iod_spec = next_seqinfo.image_type[2]
 
         # possibly apply present formatting in the series_description or protocol name
         for f in "series_description", "protocol_name":
-            s = s._replace(**{f: getattr(s, f).format(**s._asdict())})
+            curr_seqinfo = curr_seqinfo._replace(
+                **{f: getattr(curr_seqinfo, f).format(**curr_seqinfo._asdict())}
+            )
 
         template = None
         suffix = ""
         # seq = []
 
-        # figure out type of image from s.image_info -- just for checking ATM
[…]
