### 2024-06-21 2c4835e
Generalize change to other suffixes.
