### 2024-06-30 #5839 @yarikoptic
no -- current version is now 1.1.6 (was never `1.1.43`, I guess that was a typo)
- on pypi, the same URL  https://pypi.org/project/heudiconv/ leads to 1.1.6 https://pypi.org/project/heudiconv/1.1.6/
- on zenodo: it is https://zenodo.org/records/11497270 which also points to the  archive https://zenodo.org/records/11497270/files/nipy/heudiconv-v1.1.6.zip
- on github: https://github.com/nipy/heudiconv/tree/v1.1.6 pointing to release page https://github.com/nipy/heudiconv/releases/tag/v1.1.6 with sources https://github.com/nipy/heudiconv/archive/refs/tags/v1.1.6.tar.gz
