### 2024-07-03 #5839 @mvdoc
I noticed that the automatically generated crossref file has some errors in how first names/last names were parsed. I'm not sure if it matters but I just wanted to point it out.
(My last name always creates issues with automatic parsers, but I also noticed there were errors with at least another  name):

https://github.com/openjournals/joss-papers/blob/master/joss.05839/10.21105.joss.05839.crossref.xml

```
          <person_name sequence="additional"
                       contributor_role="author">
            <given_name>Matteo Visconti</given_name>
            <surname>di Oleggio Castello</surname>
            <ORCID>https://orcid.org/0000-0001-7931-5272</ORCID>
          </person_name>
...
          <person_name sequence="additional"
                       contributor_role="author">
            <given_name>John T. Wodder</given_name>
            <surname>II</surname>
          </person_name>
```
