### 2024-07-03 #5839 @danielskatz
If the names are described in the .md file as explained in https://joss.readthedocs.io/en/latest/paper.html#names, this shouldn't happen.  If they aren't, they could be fixed, and the paper could be reaccepted to fix this.
