### 2024-07-03 b0f9314
FIX split names into given-names/surname/suffix
