+++ paper.md
@@ -9,85 +9,113 @@ tags:
  - open science
  - FOSS
 authors:
- - name: Yaroslav O. Halchenko  # 626 commits, issues: opened 81 participated in 146
+ - given-names: Yaroslav O.
+   surname: Halchenko # 626 commits, issues: opened 81 participated in 146
    orcid: 0000-0003-3456-2493
    affiliation: 1
- - name: Mathias Goncalves  # 287 commits, issues: opened 16 participated in 98
+ - given-names: Mathias
+   surname: Goncalves # 287 commits, issues: opened 16 participated in 98
    orcid: 0000-0002-7252-7771
    affiliation: 2
- - name: Satrajit Ghosh  # 77 commits, issues: opened 16 participated in 62
+ - given-names: Satrajit
+   surname: Ghosh  # 77 commits, issues: opened 16 participated in 62
    orcid: 0000-0002-5312-6729
    affiliation: 3
- - name: Pablo Velasco  # 113 commits, issues: opened 6 participated in 22
+ - given-names: Pablo
+   surname: Velasco  # 113 commits, issues: opened 6 participated in 22
    orcid: 0000-0002-5749-6049
    affiliation: 4
- - name: Matteo Visconti di Oleggio Castello  # 88 commits, issues: opened 2 participated in 1
+ - given-names: Matteo
+   surname: Visconti di Oleggio Castello  # 88 commits, issues: opened 2 participated in 1
    orcid: 0000-0001-7931-5272
    affiliation: 5
- - name: Taylor Salo  # 76 commits, issues: opened 11 participated in 4
+ - given-names: Taylor
+   surname: Salo  # 76 commits, issues: opened 11 participated in 4
    orcid: 0000-0001-9813-3167
    affiliation: 6
- - name: John T. Wodder II  # 58 commits, issues: opened 2 participated in 4
+ - given-names: John T.
+   surname: Wodder # 58 commits, issues: opened 2 participated in 4
+   suffix: II
    orcid:
    affiliation: 1
- - name: Michael Hanke  # 32 commits, issues: opened 10 participated in 6
+ - given-names: Michael
+   surname: Hanke  # 32 commits, issues: opened 10 participated in 6
    orcid: 0000-0001-6398-6370
    affiliation: 7, 8
- - name: Patrick Sadil  # 18 commits, issues: opened 3 participated in 2
+ - given-names: Patrick
+   surname: Sadil  # 18 commits, issues: opened 3 participated in 2
    orcid: 0000-0003-4141-1343
    affiliation: 9
- - name: Krzysztof Jacek Gorgolewski  # 4 commits, issues: opened 7 participated in 10
+ - given-names: Krzysztof Jacek
+   surname: Gorgolewski  # 4 commits, issues: opened 7 participated in 10
    orcid: 0000-0003-3321-7583
    affiliation: 10
- - name: Horea-Ioan Ioanas   # 17 commits, issues: opened 1 participated in 1
+ - given-names: Horea-Ioan
+   surname: Ioanas   # 17 commits, issues: opened 1 participated in 1
    orcid: 0000-0001-7037-2449
    affiliation: 1
- - name: Chris Rorden  # 0 commits, issues: opened 1 participated in 17
+ - given-names: Chris
+   surname: Rorden  # 0 commits, issues: opened 1 participated in 17
    orcid: 0000-0002-7554-6142
    affiliation: 11
- - name: Timothy J. Hendrickson  # 0 commits, issues: opened 10 participated in 4
+ - given-names: Timothy J.
+   surname: Hendrickson  # 0 commits, issues: opened 10 participated in 4
    orcid: 0000-0001-6862-6526
    affiliation: 12, 13
[…]
