### 2024-07-03 #5839 @mvdoc
Thanks @danielskatz. We have fixed the .md file in https://github.com/nipy/heudiconv/pull/772.
