### 2024-07-03 #5839 @danielskatz
the track chair, @Kevin-Mattheus-Moerman, should next do a reaccept to update the metadata deposited with Crossref.
