### 2024-07-04 #5839 @danielskatz
@Kevin-Mattheus-Moerman - I see two problems in the metadata, thought neither matches the error messages
Lines 39 and 103 (empty orcid entries) should be removed.
The affiliations also aren't quoted, which might be ok when there is only one but typically fails when there are multiple.
I see that these are not new issues, so I'm unsure why they would be problems now, when they were not before, unless the logic that JOSS uses for multiple name parts is more strict that for single name fields.

perhaps @xuanxu will have thoughts as well?
