### 2024-07-04 #5839 @xuanxu
There is [a typo](https://github.com/nipy/heudiconv/blob/paper-joss/paper.md?plain=1#L70) in one of the metadata keys (`surmame` for `surname`), [this PR](https://github.com/nipy/heudiconv/pull/773) should fix the error.
