### 2024-07-08 #5839 @yarikoptic
Looks great to me,  Seems to look good for @mvdoc 's name:

https://github.com/openjournals/joss-papers/blob/master/joss.05839/10.21105.joss.05839.crossref.xml#L67

```xml
          <person_name sequence="additional"
                       contributor_role="author">
            <given_name>Matteo</given_name>
            <surname>Visconti di Oleggio Castello</surname>
            <ORCID>https://orcid.org/0000-0001-7931-5272</ORCID>
          </person_name>
```

Huge Thanks EVERYONE who was with us on this journey!
